+++ NEWS.md
@@ -1,5 +1,9 @@
 # rsconnect 0.8.30 (development version)
 
+* New `rsconnect.http.headers` and `rsconnect.http.cookies` allow you to
+  set extra arbitrary additional headers/cookies on each request (#405).
+  Their use is documented in the new `vignette("custom-http")`.
+
 * `deployApp()` now uses a stricter policy for determining whether or not
   a local package can be successfully installed on the deployment server.
   This means that you're more likely to get a clean failure prior to